examples/strategy_factory/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
duckdb==1.1.3
openai==1.54.4
httpx==0.27.2
python-dotenv==1.0.1
pandas